# evidence/ex13_capstone/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API de Agendamento Clínico",
    "description": "Pacientes, profissionais de saúde e consultas. Dados de saúde (LGPD art. 11).",
    "version": "1.0.0"
  },
  "paths": {
    "/user/signin": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Sign In",
        "description": "Login (OAuth2 Password flow). Contas admin recebem um desafio MFA em vez de\num access token; conclua em `POST /user/mfa/verify`.",
        "operationId": "sign_in_user_signin_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_sign_in_user_signin_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas (mensagem genérica)"
          },
          "429": {
            "description": "Limite estrito de tentativas excedido (veja Retry-After)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/mfa/verify": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Mfa Verify",
        "description": "Segundo fator (TOTP). Troca o `mfa_token` + código por um access token.",
        "operationId": "mfa_verify_user_mfa_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MFAVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas (mensagem genérica)"
          },
          "429": {
            "description": "Limite estrito de tentativas excedido (veja Retry-After)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/signup": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create User",
        "description": "Cria usuário interno. Restrito a administradores: não existe auto-cadastro,\ne o papel nunca é escolhido por quem não é admin (antes: mass assignment).",
        "operationId": "create_user_user_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreated"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado"
          },
          "403": {
            "description": "Perfil/escopo sem permissão para a operação"
          },
          "429": {
            "description": "Limite de requisições excedido (veja Retry-After)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/user/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Read Me",
        "operationId": "read_me_user_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado"
          },
          "403": {
            "description": "Perfil/escopo sem permissão para a operação"
          },
          "429": {
            "description": "Limite de requisições excedido (veja Retry-After)"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/user/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List Users",
        "operationId": "list_users_user__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserPublic"
                  },
                  "title": "Response List Users User  Get"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado"
          },
          "403": {
            "description": "Perfil/escopo sem permissão para a operação"
          },
          "429": {
            "description": "Limite de requisições excedido (veja Retry-After)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/{user_id}/deactivate": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Deactivate User",
        "operationId": "deactivate_user_user__user_id__deactivate_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado"
          },
          "403": {
            "description": "Perfil/escopo sem permissão para a operação"
          },
          "429": {
            "description": "Limite de requisições excedido (veja Retry-After)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_sign_in_user_signin_post": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 254,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 72,
            "title": "Password"
          },
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8,
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "scope": {
            "type": "string",
            "maxLength": 200,
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_sign_in_user_signin_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MFAVerifyRequest": {
        "properties": {
          "mfa_token": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 20,
            "title": "Mfa Token"
          },
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "title": "Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mfa_token",
          "code"
        ],
        "title": "MFAVerifyRequest"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "admin",
          "recepcionista",
          "profissional"
        ],
        "title": "Role"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "Bearer"
          },
          "expires_in": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires In"
          },
          "mfa_required": {
            "type": "boolean",
            "title": "Mfa Required",
            "default": false
          },
          "mfa_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mfa Token"
          }
        },
        "type": "object",
        "title": "TokenResponse"
      },
      "UserCreate": {
        "properties": {
          "name": {
            "type": "string",
            "pattern": "^[A-Za-zÀ-ÖØ-öø-ÿ][A-Za-zÀ-ÖØ-öø-ÿ .'\\-]{1,99}$",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 72,
            "minLength": 12,
            "title": "Password"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "email",
          "password",
          "role"
        ],
        "title": "UserCreate"
      },
      "UserCreated": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "mfa_provisioning_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mfa Provisioning Uri"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "role",
          "is_active"
        ],
        "title": "UserCreated"
      },
      "UserPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "role",
          "is_active"
        ],
        "title": "UserPublic"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/user/signin"
          }
        }
      },
      "LabClientCredentials": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "scopes": {
              "availability:read": "Consultar horários livres de um profissional"
            },
            "tokenUrl": "/oauth/token"
          }
        }
      }
    }
  }
}